Moved sunlight func to utils
