+++ src/components/Sunlight.svelte
@@ -1,35 +1,15 @@
 <script>
+	// Import utils
+	import { parseTimeInterval, getPartOfDay, getStartPosition, getEndPosition } from '@utils';
+
 	export let period;
 	export let title;
-
-	let hoursOfDay = 24;
 
 	let periodStart;
 	let periodEnd;
 	let periodRatio;
 	let periodStartPosition;
 	let periodEndPosition;
-
-	const parseTimeInterval = (interval) => {
-		if (interval === '-') return [];
-		return interval.split('-').map((v) => parseInt(v));
-	};
-
-	const getPartOfDay = (start, end) => {
-		let delta = end - start;
-		if (isNaN(delta)) return 0;
-		let ratio = (delta / hoursOfDay) * 100;
-		if (ratio === 0) return 1;
-		return Math.floor(ratio);
-	};
-
-	const getStartPosition = (start) => {
-		return (start / hoursOfDay) * 100;
-	};
-
-	const getEndPosition = (end) => {
-		return (end / hoursOfDay) * 100;
-	};
 
 	$: if (period) {
 		[periodStart, periodEnd] = parseTimeInterval(period);
@@ -40,7 +20,7 @@
 </script>
 
 <div class="flex gap-4 items-center">
-	<p class="w-[20%]">{title}</p>
+	<p class="w-[20%] m-0 font-normal">{title}</p>
 	<div class="h-3.5 flex-1 bg-deep-brown relative">
 		{#if periodStart}
 			<span class="font-light absolute -top-[140%]" style:left={`${periodStartPosition - 2}%`}
@@ -61,10 +41,3 @@
 		{/if}
 	</div>
 </div>
-
-<style>
-	p {
-		margin: 0;
-		font-weight: 400;
-	}
-</style>
